fix(oracledb_exporter): support CGO_ENABLED=0 cross-compilation
The oracledb integration imported github.com/oracle/oracle-db-appdev-monitoring/collector,
which transitively pulls in github.com/godror/godror. godror requires CGO, so
`cd collector && CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build .` failed with
"undefined: VersionInfo" and similar errors from godror's orahlp.go.

Apply the existing loki.source.journal dual-file pattern: keep pure-Go config
and helpers unconditional, move the godror-touching New() body behind
//go:build cgo, and add a //go:build !cgo stub New() that returns a clear
error so configs surface a meaningful message in no-cgo builds.

Also add CI cross-compile jobs covering linux/amd64 and linux/arm64 on
macos-15-xlarge: collector in test_mac.yml, pyroscope playground in
test_pyroscope_pr.yml.
