+++ .github/workflows/test_mac.yml
@@ -30,3 +30,23 @@ jobs:
         cache: true
     - name: Test
       run: CGO_LDFLAGS="-ld_classic $CGO_LDFLAGS" make GO_TAGS="nodocker" test
+
+  cross-compile-collector-linux-nocgo:
+    name: Cross-compile collector for linux/${{ matrix.goarch }} with CGO disabled
+    strategy:
+      matrix:
+        platform: [macos-15-xlarge]
+        goarch: [amd64, arm64]
+    runs-on: ${{ matrix.platform }}
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+      with:
+        persist-credentials: false
+    - name: Set up Go
+      uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6.4.0
+      with:
+        go-version-file: go.mod
+        cache: true
+    - name: Build
+      run: cd collector && CGO_ENABLED=0 GOOS=linux GOARCH=${{ matrix.goarch }} go build .
