+++ .github/workflows/test_pyroscope_pr.yml
@@ -27,3 +27,23 @@ jobs:
           cache: false
 
       - run: sudo make test-pyroscope
+
+  cross-compile-pyroscope-playground-linux-nocgo:
+    name: Cross-compile pyroscope playground for linux/${{ matrix.goarch }} with CGO disabled
+    strategy:
+      matrix:
+        platform: [macos-15-xlarge]
+        goarch: [amd64, arm64]
+    runs-on: ${{ matrix.platform }}
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+        with:
+          persist-credentials: false
+      - name: Set up Go
+        uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6.4.0
+        with:
+          go-version-file: go.mod
+          cache: true
+      - name: Build
+        run: CGO_ENABLED=0 GOOS=linux GOARCH=${{ matrix.goarch }} go build ./internal/component/pyroscope/util/internal/cmd/playground
